fix: address CodeRabbit review round 2 for PR #58
- transport: revert usb_pid to int32_t in 3-param overload (Hand uses -1 sentinel)
- transport: keep uint16_t in 6-param overload (TouchBoard explicit PID)
- hand_zenoh_bridge: startup dtype np.int32 → np.uint16 for control_mode
- hand_zenoh_bridge: publish loop overrun-skip (consistent with TactileBridge)
- hand_zenoh_bridge: fix liveliness DELETE race (only stop watcher if owner matches)
- touch_board.cpp: explicit destructor resets transport before member teardown
- tactile_bridge: handle handedness=-1 gracefully (tboard_unknown instead of tboard_-1)

+++ wujihandcpp/src/transport/transport.hpp
@@ -29,7 +29,7 @@ public:
 };
 
 std::unique_ptr<ITransport>
-    create_usb_transport(uint16_t usb_vid, uint16_t usb_pid, const char* serial_number);
+    create_usb_transport(uint16_t usb_vid, int32_t usb_pid, const char* serial_number);
 
 std::unique_ptr<ITransport>
     create_usb_transport(uint16_t usb_vid, uint16_t usb_pid, const char* serial_number,
